# Add Prometheus resource support