Adds `api/label.MergeIterator` for combining two sorted label sets efficiently, and a benchmark. Uses the new code to merge metric labels and resource labels.

TODO: This leaves some questions about whether the specification firmly states which of (1) resources or (2) labels should override the other.

Resolves #553.

BEFORE
```
BenchmarkMergeResource_1-8    	 2612361	       455 ns/op	     368 B/op	       6 allocs/op
BenchmarkMergeResource_2-8    	 1880131	       642 ns/op	     608 B/op	       6 allocs/op
BenchmarkMergeResource_3-8    	 1411632	       838 ns/op	     848 B/op	       6 allocs/op
BenchmarkMergeResource_4-8    	 1000000	      1060 ns/op	    1088 B/op	       6 allocs/op
BenchmarkMergeResource_6-8    	  763972	      1507 ns/op	    1568 B/op	       6 allocs/op
BenchmarkMergeResource_8-8    	  606608	      1988 ns/op	    2048 B/op	       6 allocs/op
BenchmarkMergeResource_16-8   	  232756	      5072 ns/op	    4736 B/op	       7 allocs/op
```
AFTER
```
BenchmarkMergeResource_1-8    	 3116332	       378 ns/op	     272 B/op	       4 allocs/op
BenchmarkMergeResource_2-8    	 2444085	       491 ns/op	     416 B/op	       4 allocs/op
BenchmarkMergeResource_3-8    	 1828692	       635 ns/op	     560 B/op	       4 allocs/op
BenchmarkMergeResource_4-8    	 1597401	       748 ns/op	     704 B/op	       4 allocs/op
BenchmarkMergeResource_6-8    	 1222702	       966 ns/op	     992 B/op	       4 allocs/op
BenchmarkMergeResource_8-8    	  969420	      1216 ns/op	    1280 B/op	       4 allocs/op
BenchmarkMergeResource_16-8   	  411552	      2913 ns/op	    3200 B/op	       5 allocs/op
```